changed structure and cleaned unused imports
