--- src/api/mod.rs
@@ -1,3 +1,0 @@
-pub mod signal;
-pub mod ingest;
-pub mod query_timeseries;
