+++ src/app/signal_meta/mod.rs
@@ -1,0 +1,4 @@
+pub mod route;
+pub mod controller;
+pub mod model;
+pub mod error;
