--- src/error.rs
@@ -1,10 +1,0 @@
-
-
-#[derive(thiserror::Error, Debug)]
-pub enum Error {
-    #[error("Generic{0}")]
-    Generic(String),
-
-    #[error(transparent)]
-    IO(#[from] std::io::Error),
-}
