--- src/model/mod.rs
@@ -1,3 +1,0 @@
-pub mod signal;
-pub mod ingest_packet;
-pub mod query_timeseries;
