--- src/prelude.rs
@@ -1,7 +1,0 @@
-
-
-pub use crate::error::Error;
-
-pub type Result<T> = core::result::Result<T, Error>;
-
-pub struct W<T>(pub T);
