--- src/utils/macros.rs
@@ -1,8 +1,0 @@
-macro_rules! map {
-    ($($k:expr => $v:expr),* $(,)?) => {{
-		let mut m = ::std::collections::BTreeMap::new();
-        $(m.insert($k, $v);)+
-        m
-    }};
-  }
-pub(crate) use map;
