--- src/utils/mod.rs
@@ -1,2 +1,0 @@
-use crate::prelude::*;
-pub mod macros;
